+++ src/main/java/dev/efnilite/worldserver/config/Type.java
@@ -1,9 +1,0 @@
-package dev.efnilite.worldserver.config;
-
-public enum Type {
-
-    CHAT,
-    TAB,
-    ECO
-
-}