+++ src/main/java/dev/efnilite/worldserver/hook/PlaceholderHook.java
@@ -18,12 +18,8 @@ public class PlaceholderHook {
     /**
      * If PAPI is found, translate placeholders. If not, return the given string.
      *
-     * @param   player
-     *          The player
-     *
-     * @param   string
-     *          The string
-     *
+     * @param player The player
+     * @param string The string
      * @return the string if PAPI is not found. the translated string if PAPI is found.
      */
     public static String translate(Player player, String string) {