+++ changes.md
@@ -1,3 +1,6 @@
 - Fixed switch messages doubling up on join
 - Fixed eco commands being registered when eco is disabled
 - Fixed invalid placeholders not printing an error
+- Fixed /baltop menu crashing server and colouring error
+- Fixed worlds not being prioritized over groups in join/leave messages
+- Reduced jar size